republished under new name and reconfigured JS code
republished the Pokedex app under a new name in order to eliminate the issues I was having with commits and pushing from VScode.  Now the Githubdesktop will communicate with VScode and my profile on Github.com.  Setup HTML, added CSS styling, and reconfigured the JS code from for loops to forEach() functions.