Added functions, forEach() loops and IIFE
Wrapped the Pokémon List array in an IIFE.  Changed for loops into forEach() loops.  Added an add and get all function to modify and retrieve the array from outside the IIFE.  Added a Pokémon to the array from outside the IIFE, retrieved the array and printed it.  Added parameters to the Add function to prevent the wrong objects from being added.