added filtered search and loading GIF
Adjusted add Pokémon function to stop over restricting of Pokémon list.  Adjusted API to a limit of 150 Pokémon.  Re-wrote functions to allow for keywords to work outside IIFE.  Added search bar and search functions.  Added a loading GIF and set a time limit for it.  Added styling to center the GIF, change the color and opacity of the page during the countdown.